pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AIAssistantsLib"
version = "0.1.4"
description = "A utility library for AI assistants."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name="Georgy Volkov", email="gbvolkov@gmail.com" }
]
keywords = ["AI", "assistants", "LLM", "RAG"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
requires-python = ">=3.11"

# Define dependencies as a list under [project]
dependencies = [
    "python-dotenv",
    "accelerate>=0.26.0",
    "transformers",
    "torch",
    "langchain_openai",
    "langchain_gigachat",
    "langchain_community",
    "langchainhub",
    "langchain-huggingface",
    "langchain_google_genai",
    "langchain_mistralai",
    "yandexcloud",
    "faiss-cpu",
    "nltk"
]

# Define URLs as a separate table under [project]
[project.urls]
Source = "https://github.com/gbvolkov/AIAssistantsLibPackage"
Tracker = "https://github.com/gbvolkov/AIAssistantsLibPackage/issues"
